chore: separate db connection test from config validation
